+++ app/etl/etl-updatedb/src/main/java/fr/gouv/monprojetsup/data/etl/loaders/DataSources.java
@@ -44,7 +44,7 @@ public class DataSources {
     public static final String RESUMES_MPS_PATH = "mps/Tableau_resume_descriptifs_formations.csv";
     public static final String DOMAINES_MPS_PATH = "mps/domaines_pro_MPS_-_Thematiques.tsv";
     public static final String INTERETS_GROUPES_PATH = "mps/centres_d_interets_MPS_-_Feuille_1.tsv";
-    public static final String PSUP_TO_IDEO_CORRESPONDANCE_PATH = "onisep_ideo_hotline/T_livre_a_Parcoursup_2024_01_26_HG.csv";
+    public static final String PSUP_TO_IDEO_CORRESPONDANCE_PATH = "onisep_ideo_hotline/liensPSUP_ONISEP_2024_09_06_V1_HG.csv";
     public static final String MPS_HERITAGES_PATH = "mps/MPS_heritages.csv";
     public static final String MPS_HERITAGES_HERITIER_HEADER = "mps_heritier";
     public static final String MPS_HERITAGES_LEGATAIRES_HEADER = "mps_legataire";
