+++ app/etl/etl-updatedb/src/test/kotlin/fr/gouv/monprojetsup/data/etl/loaders/EdgesTest.kt
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 
-@SpringBootTest(classes = [LoadersTest::class, DataSources::class])
+@SpringBootTest(classes = [DataSources::class])
 class EdgesTest {
 
     @Autowired
