--- app/suggestions/suggestions-export/src/main/kotlin/fr/gouv/monprojetsup/suggestions/export/JpaConfig.kt
@@ -1,6 +1,0 @@
-package fr.gouv.monprojetsup.suggestions.export
-
-import org.springframework.boot.autoconfigure.domain.EntityScan
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
